[ie/youtube] Move registry from globals to pot module
