+++ test/test_pot/test_pot_builtin_webpospec.py
@@ -11,7 +11,7 @@ from yt_dlp.extractor.youtube.pot.provider import (
 from yt_dlp.version import __version__
 
 from yt_dlp.extractor.youtube.pot._builtin.webpo_cachespec import WebPoPCSP
-from yt_dlp.globals import _pot_pcs_providers
+from yt_dlp.extractor.youtube.pot._registry import _pot_pcs_providers
 
 
 @pytest.fixture()
