+++ yt_dlp/extractor/youtube/pot/cache.py
@@ -13,8 +13,12 @@ from yt_dlp.extractor.youtube.pot._provider import (
     register_preference_generic,
     register_provider_generic,
 )
+from yt_dlp.extractor.youtube.pot._registry import (
+    _pot_cache_provider_preferences,
+    _pot_cache_providers,
+    _pot_pcs_providers,
+)
 from yt_dlp.extractor.youtube.pot.provider import PoTokenRequest
-from yt_dlp.globals import _pot_cache_provider_preferences, _pot_cache_providers, _pot_pcs_providers
 
 
 class PoTokenCacheProviderError(IEContentProviderError):
